Gestion de la création d'un show

+++ src/main/java/eu/frezilla/fundraising_progress/entity/ShowProjectLink.java
@@ -47,7 +47,7 @@ public class ShowProjectLink implements Comparable<ShowProjectLink>, Serializabl
 
         @Column(name = "show_id", nullable = false)
         @EqualsAndHashCode.Include
-        private Integer showId;
+        private Long showId;
 
         @Override
         public int compareTo(Id id) {
